requirements.txt
alembic==1.8.1
autopep8==1.6.0
click==8.1.3
Flask==2.1.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
SQLAlchemy==1.4.39
toml==0.10.2
Werkzeug==2.2.0
zipp==3.8.1